Annotate print_message() function with PRINTF macro and fix warnings
The PRINTF macro (defined in lib/macros.h) expands to a GCC/Clang
function attribute that allows the compiler to show warnings for
mismatched printf specifers.

+++ nokogumbo-import/gumbo-parser/src/error.h
@@ -147,7 +147,7 @@ typedef struct GumboInternalError {
     // * GUMBO_ERR_UTF8_TRUNCATED
     // * GUMBO_ERR_NUMERIC_CHAR_REF_WITHOUT_SEMICOLON
     // * GUMBO_ERR_NUMERIC_CHAR_REF_INVALID
-    uint64_t codepoint;
+    uint32_t codepoint;
 
     // Tokenizer errors.
     GumboTokenizerError tokenizer;
